+++ src/pages/RECIPES/CreateRecipes/RecipeForm.jsx
@@ -16,49 +16,20 @@ export const RecipeForm = ({
     handleSubmit,
     handleImageChange,
 }) => {
-    const handleInputIngre = (e, index) => {
-        const novosValores = [...formData.ingredients]
-        novosValores[index] = e.target.value
-        setFormData({ ...formData, ingredients: novosValores })
-    }
-    const adicionarIngre = () => {
-        setFormData({ ...formData, ingredients: [...formData.ingredients, ''] })
-    }
-    const removerIngre = (index) => {
-        const novosValores = [...formData.ingredients]
-        novosValores.splice(index, 1)
-        setFormData({ ...formData, ingredients: novosValores })
-    }
-    const handleInputModPreps = (e, index) => {
-        const novosValores = [...formData.modPreps]
-        novosValores[index] = e.target.value
-        setFormData({ ...formData, modPreps: novosValores })
-    }
-    const adicionarModPreps = () => {
-        setFormData({ ...formData, modPreps: [...formData.modPreps, ''] })
-    }
-    const removerModPreps = (index) => {
-        const novosValores = [...formData?.modPreps]
-        novosValores.splice(index, 1)
-        setFormData({ ...formData, modPreps: novosValores })
-    }
-
-    
+  
     return (
         <>
             <MyTypography title={"Criação de receitas"} variant="h6" />
             <HeaderCreateRecipes
-                handleInputIngre={handleInputIngre}
-                adicionarIngre={adicionarIngre}
-                removerIngre={removerIngre}
+                formData={formData}
+                setFormData={setFormData}
                 handleInputChangesCreateRecipes={handleInputChangesCreateRecipes}
             />
             <MyTypography title={"Etapas de preparo"} variant="h6" gutterBottom mt={3} mb={1} />
             <Grid container spacing={3}>
                 <StepsCreateRecipes
-                    handleInputModPreps={handleInputModPreps}
-                    adicionarModPreps={adicionarModPreps}
-                    removerModPreps={removerModPreps}
+                    formData={formData}
+                    setFormData={setFormData}
                 />
             </Grid>
             <MyTypography title={"Categorias e Nutrição"} variant="h6" gutterBottom mt={3} />
